+++ paddle/fluid/operators/ngraph/CMakeLists.txt
@@ -1,5 +1,6 @@
 if(WITH_NGRAPH)
   cc_library(ngraph_bridge SRCS ngraph_bridge.cc DEPS operator framework_proto ngraph)
-  cc_library(ngraph_operator SRCS ngraph_operator.cc DEPS ngraph_bridge operator op_info device_context tensor scope glog
+  cc_library(ngraph_engine SRCS ngraph_engine.cc DEPS ngraph_bridge)
+  cc_library(ngraph_operator SRCS ngraph_operator.cc DEPS ngraph_engine operator op_info device_context tensor scope glog
              shape_inference data_transform lod_tensor profiler)
 endif(WITH_NGRAPH)
